+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [0.11]
+
+### Bugfix
+
+- Some potential data races fixed (see the `init_TestU01_internals` function
+  in `src\testu01_callbacls.cpp` for details). The bug was reported by
+  [skeeto](https://www.reddit.com/r/RNG/comments/1vwnr1l/comment/p5invw8).
+- Binary stdin/stdout mode now correctly works in `src/testu01th_demo.cpp` and
+  `src/testu01th_pipes.cpp`. It allows to send PRNG output to TestU01 through
+  stdin/stdout (just as in PractRand) but only in one-threaded mode.
+- `testu01th_pipes`: obsolete command line arguments documentation fixed.
+
+### Changed
+
+- CMakeLists.txt refactoring.
+
 ## [0.10] 2026-08-24
 
 An initial pre-release.